+++ bailian-20231229/src/Models/AddConnectorRequest/fileConnectorConfig.php
@@ -1,0 +1,76 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Bailian\V20231229\Models\AddConnectorRequest;
+
+use AlibabaCloud\Dara\Model;
+
+class fileConnectorConfig extends Model
+{
+    /**
+     * @var string
+     */
+    public $bucketName;
+
+    /**
+     * @var string
+     */
+    public $regionId;
+
+    /**
+     * @var string
+     */
+    public $storageType;
+    protected $_name = [
+        'bucketName' => 'BucketName',
+        'regionId' => 'RegionId',
+        'storageType' => 'StorageType',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->bucketName) {
+            $res['BucketName'] = $this->bucketName;
+        }
+
+        if (null !== $this->regionId) {
+            $res['RegionId'] = $this->regionId;
+        }
+
+        if (null !== $this->storageType) {
+            $res['StorageType'] = $this->storageType;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['BucketName'])) {
+            $model->bucketName = $map['BucketName'];
+        }
+
+        if (isset($map['RegionId'])) {
+            $model->regionId = $map['RegionId'];
+        }
+
+        if (isset($map['StorageType'])) {
+            $model->storageType = $map['StorageType'];
+        }
+
+        return $model;
+    }
+}