+++ bailian-20231229/src/Models/GetConnectorRequest.php
@@ -1,0 +1,62 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Bailian\V20231229\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class GetConnectorRequest extends Model
+{
+    /**
+     * @var string
+     */
+    public $connectorId;
+
+    /**
+     * @var string
+     */
+    public $connectorName;
+    protected $_name = [
+        'connectorId' => 'ConnectorId',
+        'connectorName' => 'ConnectorName',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->connectorId) {
+            $res['ConnectorId'] = $this->connectorId;
+        }
+
+        if (null !== $this->connectorName) {
+            $res['ConnectorName'] = $this->connectorName;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['ConnectorId'])) {
+            $model->connectorId = $map['ConnectorId'];
+        }
+
+        if (isset($map['ConnectorName'])) {
+            $model->connectorName = $map['ConnectorName'];
+        }
+
+        return $model;
+    }
+}