+++ bailian-20231229/src/Models/GetConnectorResponseBody/data.php
@@ -1,0 +1,90 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Bailian\V20231229\Models\GetConnectorResponseBody;
+
+use AlibabaCloud\Dara\Model;
+
+class data extends Model
+{
+    /**
+     * @var string
+     */
+    public $connectorId;
+
+    /**
+     * @var string
+     */
+    public $connectorName;
+
+    /**
+     * @var string
+     */
+    public $connectorType;
+
+    /**
+     * @var string
+     */
+    public $description;
+    protected $_name = [
+        'connectorId' => 'ConnectorId',
+        'connectorName' => 'ConnectorName',
+        'connectorType' => 'ConnectorType',
+        'description' => 'Description',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->connectorId) {
+            $res['ConnectorId'] = $this->connectorId;
+        }
+
+        if (null !== $this->connectorName) {
+            $res['ConnectorName'] = $this->connectorName;
+        }
+
+        if (null !== $this->connectorType) {
+            $res['ConnectorType'] = $this->connectorType;
+        }
+
+        if (null !== $this->description) {
+            $res['Description'] = $this->description;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['ConnectorId'])) {
+            $model->connectorId = $map['ConnectorId'];
+        }
+
+        if (isset($map['ConnectorName'])) {
+            $model->connectorName = $map['ConnectorName'];
+        }
+
+        if (isset($map['ConnectorType'])) {
+            $model->connectorType = $map['ConnectorType'];
+        }
+
+        if (isset($map['Description'])) {
+            $model->description = $map['Description'];
+        }
+
+        return $model;
+    }
+}